Refactor tab manager into dedicated module
